requirements.txt
flask-cors
ffprobe==0.5
Flask==1.1.1
Flask-RESTful==0.3.8
google-cloud-speech
pydub==0.23.1
Wave==0.0.2
uWSGI
Werkzeug==1.0.0
ffmpeg-python
ffmpeg
Flask-Mail
Flask-JWT-Extended
gunicorn
pandas
numpy
requests
scipy
